First some refactor!
